Let FFmpeg.ExtractCover unconditionally extract & return a new cover
Previously, FFmpeg.ExtractCover would skip extracting a new cover if
the given Track has a local file assigned to it.

Now, it will proceed with the extraction regardless of the Track's
state, with the expectation that the calling code will deal with
skipping the extraction and using a local cover accordingly.