Download *only* the audio when all Tracks have local existing covers
Gunloader downloads both the video & audio from a YouTube URL, for the
sake of using frames within the video as covers for each Track.

When all Tracks have local images assigned as covers to them, it's not
necessary to download the video from YouTube -- instead, we could only
extract the audio.

This commit teaches Gunloader to download *only* the audio stream from
YouTube when all of the Album Tracks have local images assigned to them.