Implement rudimentary GUI for Gunloader
This commit introduces a bare minimum Windows GUI for Gunloader to get
the ball rolling. Due to a current lack of time, the GUI is incredibly
spartan and frankly rather nasty as far as the code goes.

Down the line, significant improvements will hopefully be carried out
both in the functionality and overall design of the blessed thing.