Have the wizard continue only when the records file is actually edited
We do this by checking if the hash of the records file has changed.

This is amongst the infinite ways we can cleverly ensure that users
don't make mistakes, or skip any steps, when using our programs.

They'll still figure out a way of doing so.

This is a perpetual war.