Develop and integrate the Web GUI prototype into the project
This not-so-tiny commit introduces all the foundational additions to the
Web GUI with the intent of making it an usable platform.

It also introduces the changes to the CLI wizard to take an ID from the
Web GUI and utilise it, accordingly.

This is in early stages stuff, so proper documentation and improvements
will appear down the line!