Use the latest Entry's end time as the default start time
This commit introduces the necessary logic, tests and view updates to
use a Record's latest Entry's end time as the default start time.

This introduction speeds up the process of sequentially adding the
timings for each song in a video.

Down the line, some client-side code should be implemented to quickly
assign the video player's current position as the starting/ending time.

However, the plan is to use as little JS as possible until absolutely
necessary.