Introduce the 'advanced mode' importing mechanism
This commit introduces the necessary logic and UI additions/changes to
allow Tracks to be imported from plain text.

This feature is similar to the concept of using a Records text file in
the CLI. It might help a lot with the efficiency of importing songs!

For example, one could copy and paste the track titles & timings in the
video's description and format it accordingly, rather than manually add
a track using the form.

------------------------------------------------------------------------

The import data MUST adhere to the following specification:

-   One track per line; use multiple lines for multiple tracks
-   Delimited by | (each attribute is separated by '|')
-   Start & End times are in the HH:MM:SS format
-   Artists are separated by ;
-   Number, Title, Start & End times are REQUIRED
-   Album, Genre & Artist are OPTIONAL

The order of the attributes MUST be as follows:

1.  number
2.  title
3.  start time (hh:mm:ss)
4.  end time (hh:mm:ss)
5.  album
6.  genre
7.  artists

------------------------------------------------------------------------

Since there are now two forms - one simple and one advanced - they have
been extracted into their separate views.